# docs-site/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Road Freight TMS API",
    "version": "0.2.0",
    "description": "Multi-tenant Transportation Management System / Hệ thống quản lý vận tải đa tổ chức: RBAC + DAC, route optimization, live GPS tracking, AI Agent, chatbot, reporting.",
    "contact": {
      "name": "Trịnh Công Đức",
      "email": "ductuyetvoi@gmail.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local dev"
    },
    {
      "url": "/",
      "description": "Same origin (production)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/payroll/drivers": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Driver payroll and commission by month / Bảng lương và hoa hồng tài xế theo tháng",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "string",
              "example": "2026-05"
            }
          },
          {
            "in": "query",
            "name": "driverId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lương từng tài xế + tổng"
          }
        }
      }
    },
...
  }
}